pyproject.toml
[tool.poetry]
name = "geo-rect"
version = "0.1.0"
description = "Sistema Híbrido de Validación Geoespacial con CV + ML"
authors = ["Arthur Zizumbo"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# Deep Learning - PyTorch with CUDA 13.0
torch = {url = "https://download.pytorch.org/whl/cu130/torch-2.9.1%2Bcu130-cp312-cp312-win_amd64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu130/torchvision-0.24.1%2Bcu130-cp312-cp312-win_amd64.whl"}

# Computer Vision
opencv-python = "^4.12.0.88"
kornia = "^0.8.2"
pillow = "^12.0.0"
scikit-image = "^0.25.2"

# Geospatial
geopandas = "^1.1.1"
shapely = "^2.0.6"
rasterio = "^1.4.3"
pyproj = "^3.7.2"
fiona = "^1.10.1"

# Machine Learning
xgboost = "^3.1.2"
lightgbm = "^4.6.0"
scikit-learn = "^1.7.2"
optuna = "^4.6.0"
shap = "^0.50.0"

# Embeddings & Transformers
transformers = "4.56.2"
timm = "^1.0.11"

# Data Processing
polars = "^1.35.1"
pyarrow = "^21.0.0"

# MLOps
dvc =  "^3.63.0"
mlflow = "^3.6.0"

# API
fastapi = "^0.122.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"

# Utilities
python-dotenv = "^1.2.1"
httpx = "^0.28.1"
diskcache = "^5.6.3"
tqdm = "^4.67.0"
loguru = "^0.7.2"
joblib = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"
ruff = "^0.14.6"
mypy = "^1.18.2"
pre-commit = "^4.5.0"
ipykernel = "^7.1.0"
jupyter = "^1.1.1"
nbformat = "^5.10.4"

# [[tool.poetry.source]]
# name = "pytorch-cu130"
# url = "https://download.pytorch.org/whl/cu130"
# priority = "explicit"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
